add env and config file handler
