+++ src/utils/database.py
@@ -1,12 +1,10 @@
 import os
 from sqlalchemy.orm import sessionmaker
 from sqlalchemy import create_engine
-from dotenv import load_dotenv
-
-load_dotenv()
+from src.utils.env_handler import DATABASE_URL as ENV_DATABASE_URL
 
 # Engine and Session setup (defaults to local Postgres)
-DATABASE_URL = os.getenv("DATABASE_URL")
+DATABASE_URL = ENV_DATABASE_URL
 
 # Prefer psycopg v3 driver if not specified
 if DATABASE_URL.startswith("postgresql://") and "+psycopg" not in DATABASE_URL:
